IGNITE-18729: Deployment Code functionality
Fix comments
